refactor: extract `AdministrationConfiguration` out of `Configuration`
