+++ src/main/kotlin/com/github/shirleh/administration/AdministrationConfiguration.kt
@@ -1,7 +1,7 @@
-package com.github.shirleh
+package com.github.shirleh.administration
 
 import com.typesafe.config.Config
 
-class Configuration(config: Config) {
+class AdministrationConfiguration(config: Config) {
     val superuserRoleId: Long = config.getLong("discord.superuserRoleId")
 }
