+++ src/main/kotlin/com/github/shirleh/administration/Module.kt
@@ -3,5 +3,7 @@ package com.github.shirleh.administration
 import org.koin.dsl.module
 
 val administrationModule = module {
+    single(createdAtStart = true) { AdministrationConfiguration(get()) }
+
     single<ChannelRepository> { ChannelRepositoryImpl() }
 }
